xcode: auto-select Xcode prefix more carefully
Previously, if the active developer directory (set via `xcode-select`)
was set to the CLT (Command Line Tools), Homebrew would ask Spotlight
for an Xcode installation. The first match was picked, even if it was an
outdated version located on an external (or any other non-system) drive.

Now only an installation in one of the default prefixes is considered
and is only picked if the HOMEBREW_XCODE_AUTOSELECT environment variable
is set. Furthermore, there's a method to check whether the prefix was
auto-selected, i.e. it differs from the one returned by `xcode-select`
(when pointing at the CLT).

Also fix caching in the case where `Xcode.prefix` returns `nil`.