os/mac/cctools_mach: improve 'otool' parsing
Avoid `otool` failure for file names with parentheses by adding the `-m`
switch to the invocation. Always parse all architectures of a fat binary
to also see dependencies that only apply to a single architecture. Also
remove duplicates from the list.

(This also fixes the rare case where parsing would create unexpected
results if the parsed file does not contain an architecture that matches
the host architecture, causing `otool` to operate as if `-arch all` was
passed.)