Encode aliases of internal commands as files
This allows removing the duplicated alias-command mapping that currently
cannot be shared by Ruby and shell code (but is required by both). Using
plain files instead of symlinks allows them to carry extra information
(currently only used in `instal.alias` for an explanatory comment).

The only machine-readable information is the first line that is expected
to be the canonical command name corresponding to the alias.