exceptions: don't fail on GitHub API error
Failed builds will raise a `BuildError` exception. Dumping it can itself
cause an exception when querying the GitHub API, e.g. because the query
is malformed, the tap repository isn't hosted on GitHub, or because of
rate limiting (already previously handled) or authentication issues.

The build failure is the most important information, while the issue
list is just bonus information, thus a failure to obtain it shouldn't
cause Homebrew to stumble when printing the primary information. Thus
rescue *all* GitHub API errors and print them as a mere warning.